+++ AndroidApp/app/src/main/res/layout/activity_confirmation_add_product.xml
@@ -1268,6 +1268,7 @@
                     </LinearLayout>
                 </LinearLayout>
 
+                <!-- extra videos fee -->
                 <LinearLayout
                     android:id="@+id/containerVideoExtraFee"
                     android:layout_width="match_parent"
@@ -1289,7 +1290,6 @@
                     <LinearLayout
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content">
-
 
                         <TextView
                             android:id="@+id/tvVideoExtraFee"
@@ -1470,8 +1470,8 @@
                     android:layout_marginTop="@dimen/_10sdp"
                     android:background="@drawable/edittext_bg"
                     android:orientation="horizontal"
-                    android:visibility="visible"
-                    android:padding="@dimen/_5sdp">
+                    android:padding="@dimen/_5sdp"
+                    android:visibility="visible">
 
                     <TextView
                         android:id="@+id/tvMyPointsPayment"
