+++ AndroidApp/app/src/main/res/layout/activity_listing_duration.xml
@@ -26,161 +26,6 @@
             android:orientation="vertical">
 
             <LinearLayout
-                android:id="@+id/contianerClosingOption"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical">
-
-                <LinearLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:gravity="center"
-                    android:orientation="horizontal">
-
-
-                    <TextView
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:fontFamily="@font/madani_arabic_medium"
-                        android:text="@string/close_time"
-                        android:textColor="@color/textColor"
-                        android:textSize="@dimen/_14sdp" />
-
-                    <TextView
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_marginStart="@dimen/_2sdp"
-                        android:fontFamily="@font/madani_arabic_medium"
-                        android:text="*"
-                        android:textColor="@color/Red"
-                        android:textSize="@dimen/_14sdp" />
-
-                </LinearLayout>
-
-
-                <LinearLayout
-                    android:id="@+id/closingAuctionOption1"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="@dimen/_5sdp"
-                    android:background="@drawable/field_selection_border"
-                    android:orientation="vertical"
-                    android:padding="@dimen/_10sdp">
-
-
-                    <RelativeLayout
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content">
-
-                        <TextView
-                            android:id="@+id/FixedLength"
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:layout_centerVertical="true"
-                            android:fontFamily="@font/madani_arabic_medium"
-                            android:text="@string/FixedLength"
-                            android:textColor="@drawable/field_selection_color"
-                            android:textSize="@dimen/_10sdp" />
-
-                        <RadioButton
-                            android:id="@+id/btnRadioClosingAuctionOption1"
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:layout_alignParentEnd="true"
-                            android:layout_centerVertical="true" />
-                    </RelativeLayout>
-
-
-                    <TextView
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:fontFamily="@font/madani_arabic_light"
-                        android:text="@string/without_additional_expenses"
-                        android:textColor="@color/textColor"
-                        android:textSize="@dimen/_10sdp" />
-
-
-                    <androidx.recyclerview.widget.RecyclerView
-                        android:id="@+id/rvClosingTimeListOption1"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:background="@color/white"
-                        android:orientation="horizontal"
-                        android:padding="@dimen/_5sdp"
-                        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                        tools:listitem="@layout/selection_item" />
-
-
-                </LinearLayout>
-
-
-                <LinearLayout
-                    android:id="@+id/closingAuctionOption2"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="@dimen/_15sdp"
-                    android:background="@drawable/field_selection_border"
-                    android:orientation="vertical">
-
-
-                    <RelativeLayout
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:paddingStart="@dimen/_10sdp"
-                        android:paddingEnd="@dimen/_10sdp">
-
-                        <TextView
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:layout_centerVertical="true"
-                            android:fontFamily="@font/madani_arabic_medium"
-                            android:text="@string/set_your_own_time_and_time"
-                            android:textColor="@drawable/field_selection_color"
-                            android:textSize="@dimen/_10sdp" />
-
-                        <RadioButton
-                            android:id="@+id/btnRadioClosingAuctionOption2"
-                            android:layout_width="wrap_content"
-                            android:layout_height="wrap_content"
-                            android:layout_alignParentEnd="true"
-                            android:layout_centerVertical="true" />
-                    </RelativeLayout>
-
-
-                    <TextView
-                        android:id="@+id/tvPriceCustomClosingAuctionOption2"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:fontFamily="@font/madani_arabic_light"
-                        android:paddingStart="@dimen/_10sdp"
-                        android:paddingEnd="@dimen/_10sdp"
-                        android:text="@string/Rayal"
-                        android:textColor="@color/textColor"
-                        android:textSize="@dimen/_10sdp" />
-
-                    <View
-                        android:layout_width="match_parent"
-                        android:layout_height="@dimen/_1sdp"
-                        android:layout_marginTop="@dimen/_10sdp"
-                        android:background="@color/line" />
-
-                    <TextView
-                        android:id="@+id/tvClosingAuctionCustomDataOption2"
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content"
-                        android:layout_margin="@dimen/_10sdp"
-                        android:background="@drawable/edittext_bg"
-                        android:drawablePadding="@dimen/_15sdp"
-                        android:gravity="center_vertical"
-                        android:hint="@string/SelectTime"
-                        android:padding="@dimen/_10sdp"
-                        app:drawableStartCompat="@drawable/ic_calendar" />
-
-                </LinearLayout>
-            </LinearLayout>
-
-
-            <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="@dimen/_10sdp"
